requirements.txt
# Core dependencies
python-dateutil>=2.8.2
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0

# Document parsing
pdfplumber>=0.10.0
python-docx>=1.0.0

# Fuzzy matching
rapidfuzz>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
